refactor: Move GitCredentialMessage ops to impl
